+++ packages/main/src/webComponents/CustomListItem/CustomListItem.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { CustomListItem } from '@ui5/webcomponents-react/lib/CustomListItem';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('CustomListItem', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<CustomListItem />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<CustomListItem />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });