+++ packages/main/src/webComponents/DateRangePicker/DateRangePicker.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { DateRangePicker } from '@ui5/webcomponents-react/lib/DateRangePicker';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('DateRangePicker', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<DateRangePicker />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<DateRangePicker />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });