+++ packages/main/src/webComponents/Dialog/Dialog.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Dialog } from '@ui5/webcomponents-react/lib/Dialog';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Dialog', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Dialog />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Dialog />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });