+++ packages/main/src/webComponents/DurationPicker/DurationPicker.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { DurationPicker } from '@ui5/webcomponents-react/lib/DurationPicker';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('DurationPicker', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<DurationPicker />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<DurationPicker />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });