+++ packages/main/src/webComponents/GroupHeaderListItem/GroupHeaderListItem.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { GroupHeaderListItem } from '@ui5/webcomponents-react/lib/GroupHeaderListItem';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('GroupHeaderListItem', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<GroupHeaderListItem />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<GroupHeaderListItem />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });