+++ packages/main/src/webComponents/Label/Label.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Label } from '@ui5/webcomponents-react/lib/Label';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Label', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Label />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Label />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });