+++ packages/main/src/webComponents/List/List.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { List } from '@ui5/webcomponents-react/lib/List';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('List', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<List />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<List />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });