+++ packages/main/src/webComponents/MultiComboBoxItem/MultiComboBoxItem.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { MultiComboBoxItem } from '@ui5/webcomponents-react/lib/MultiComboBoxItem';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('MultiComboBoxItem', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<MultiComboBoxItem />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<MultiComboBoxItem />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });