+++ packages/main/src/webComponents/NotificationListItem/NotificationListItem.test.tsx
@@ -1,10 +1,19 @@
+import { render } from '@shared/tests';
 import { NotificationListItem } from '@ui5/webcomponents-react/lib/NotificationListItem';
-import { mount } from 'enzyme';
+import { NotificationOverflowAction } from '@ui5/webcomponents-react/lib/NotificationOverflowAction';
+import { Priority } from '@ui5/webcomponents-react/lib/Priority';
 import React from 'react';
 
 describe('NotificationListItem', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<NotificationListItem />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(
+      <NotificationListItem
+        heading="My Notification"
+        showClose
+        priority={Priority.High}
+        actions={<NotificationOverflowAction text="My Action" />}
+      />
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
 });