+++ packages/main/src/webComponents/Option/Option.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Option } from '@ui5/webcomponents-react/lib/Option';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Option', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Option />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Option />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });