+++ packages/main/src/webComponents/Panel/Panel.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Panel } from '@ui5/webcomponents-react/lib/Panel';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Panel', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Panel />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Panel />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });