+++ packages/main/src/webComponents/SideNavigationItem/SideNavigationItem.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { SideNavigationItem } from '@ui5/webcomponents-react/lib/SideNavigationItem';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('SideNavigationItem', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<SideNavigationItem />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<SideNavigationItem />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });