+++ packages/main/src/webComponents/Switch/Switch.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Switch } from '@ui5/webcomponents-react/lib/Switch';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Switch', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Switch />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Switch />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });