+++ packages/main/src/webComponents/Table/Table.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Table } from '@ui5/webcomponents-react/lib/Table';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Table', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Table />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Table />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });