+++ packages/main/src/webComponents/TableCell/TableCell.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { TableCell } from '@ui5/webcomponents-react/lib/TableCell';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('TableCell', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<TableCell />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<TableCell />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });