+++ packages/main/src/webComponents/TextArea/TextArea.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { TextArea } from '@ui5/webcomponents-react/lib/TextArea';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('TextArea', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<TextArea />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<TextArea />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });