+++ packages/main/src/webComponents/ToggleButton/ToggleButton.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { ToggleButton } from '@ui5/webcomponents-react/lib/ToggleButton';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('ToggleButton', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<ToggleButton />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<ToggleButton />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });