+++ packages/main/src/webComponents/TreeItem/TreeItem.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { TreeItem } from '@ui5/webcomponents-react/lib/TreeItem';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('TreeItem', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<TreeItem />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<TreeItem />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });